(PHP 4 >= 4.1.0, PHP 5, PHP 7) $_POST -- $HTTP_POST_VARS [deprecated] — HTTP POST variables
(PHP 4, PHP 5)
mysql_query — Send a MySQL query
Warning
This extension was deprecated in PHP 5.5.0, and it was removed in PHP 7.0.0.
Instead, the MySQLi or PDO_MySQL extension should be used.

(PHP 4, PHP 5)
mysql_num_rows — Get number of rows in result
Warning
This extension was deprecated in PHP 5.5.0, and it was removed in PHP 7.0.0.
Instead, the MySQLi or PDO_MySQL extension should be used.

(PHP 4, PHP 5 < 5.4.0)
session_register — Register one or more global variables with the current session
Warning
This function has been DEPRECATED as of PHP 5.3.0 and REMOVED as of PHP 5.4.0.