(PHP 4 >= 4.1.0, PHP 5, PHP 7) $_GET -- $HTTP_GET_VARS [deprecated] — HTTP GET variables
(PHP 4, PHP 5)
mysql_pconnect — Open a persistent connection to a MySQL server
Warning
This extension was deprecated in PHP 5.5.0, and it was removed in PHP 7.0.0.
Instead, the MySQLi or PDO_MySQL extension should be used. See also MySQL: choosing an API guide and related FAQ for more information. Alternatives to this function include:
mysqli_connect() with p: host prefix

(PHP 4, PHP 5)
mysql_select_db — Select a MySQL database
Warning
This extension was deprecated in PHP 5.5.0, and it was removed in PHP 7.0.0.
Instead, the MySQLi or PDO_MySQL extension should be used.

(PHP 4, PHP 5)
mysql_query — Send a MySQL query

Warning
This extension was deprecated in PHP 5.5.0, and it was removed in PHP 7.0.0.
Instead, the MySQLi or PDO_MySQL extension should be used.
(PHP 4, PHP 5)
mysql_num_rows — Get number of rows in result
Warning
This extension was deprecated in PHP 5.5.0, and it was removed in PHP 7.0.0.
Instead, the MySQLi or PDO_MySQL extension should be used.

(PHP 4, PHP 5)
mysql_fetch_array — Fetch a result row as an associative array, a numeric array, or both

Warning
This extension was deprecated in PHP 5.5.0, and it was removed in PHP 7.0.0.
Instead, the MySQLi or PDO_MySQL extension should be used.